feat: add ErrorStruct to moonlink_error crate
