+++ src/moonlink_error/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "moonlink_error"
+version = { workspace = true }
+edition = { workspace = true }
+license = { workspace = true }
+
+[dependencies]
+anyhow = { workspace = true }
+thiserror = { workspace = true }
