+++ src/moonlink_error/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod error;
+
+pub use error::{ErrorStatus, ErrorStruct};
