+++ src/moonlink_rpc/Cargo.toml
@@ -7,6 +7,7 @@ license = "MIT"
 [dependencies]
 bincode = { workspace = true }
 moonlink = { path = "../moonlink" }
+moonlink_error = { path = "../moonlink_error" }
 paste = { workspace = true }
 serde = { workspace = true }
 thiserror = { workspace = true }
